fix: corrige erro de null nas entidades do banco de dados

+++ src/main/java/com/estoque/CadastroDeEstoque/DTO/ProdutosDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,7 +15,7 @@ public class ProdutosDTO {
 
     private String nome;
 
-    private float preco;
+    private BigDecimal preco;
 
     private int quantidade;
